Bugsolving: First testing finds many small issues
Intent with new NEW_TIME_FRAME
Change functions to suspend
Use camelCase for names
Add companion functions like createAndLoad
Improve selectTransactionId
Improve updateTotal
viewModelScope.launch for suspend functions
Improve billPrinter, kitchenPrinter
GRPC messages sort, start with Capital
poe_loading Loading text when loading items over grpc

+++ microfood/app/src/main/java/com/hha/framework/CClientItemsHandler.kt
@@ -37,7 +37,8 @@ class CclientItemsHandler {
         mTimeFrameIndex = timeFrameIndex
         mTransactionId = transactionId
         val itemsDb = mItemsDb.selectTransactionId(
-            transactionId, EItemSort.toItemSort(sort), timeFrameIndex.index.toInt(), -1)
+            transactionId, EItemSort.toItemSort(sort),
+            timeFrameIndex.index, -1)
         mSubItems = if (CFG.getOption("entry_show_deleted"))
             ESubItems.ORDER_ITEMS_AND_DELETED_AND_EXTRA
         else ESubItems.ORDER_ITEMS_AND_EXTRA
